add requested changes for pr
