+++ package.json
@@ -17,11 +17,15 @@
     "@fontsource/inter": "^5.1.1",
     "@mui/icons-material": "^6.4.0",
     "@mui/joy": "^5.0.0-beta.51",
+    "@mui/lab": "^6.0.0-beta.26",
+    "@mui/material": "^6.4.3",
+    "@mui/x-date-pickers": "^7.26.0",
     "@reduxjs/toolkit": "^2.5.0",
     "@stomp/stompjs": "^7.0.0",
     "@tanstack/react-query": "^5.64.2",
     "@uidotdev/usehooks": "^2.4.1",
     "axios": "^1.7.9",
+    "dayjs": "^1.11.13",
     "global": "^4.4.0",
     "i18next": "^24.2.2",
     "jwt-decode": "^4.0.0",
