+++ src/customTypes/bffTypes.tsx
@@ -1,12 +1,12 @@
 export interface Athlete {
-  id: number;
+  id?: number;
   first_name: string;
   last_name: string;
   email: string;
   birthdate: string;
   gender: string;
-  total_medal: MedalType;
-  swimming_certificate: boolean;
+  total_medal?: MedalType;
+  swimming_certificate?: boolean;
 }
 
 export enum MedalType {
