+++ src/i18n/de-DE/translation.json
@@ -32,15 +32,14 @@
       "header": "Athlet*innen",
       "createButton": "erstelle Athlet*in"
     },
-
     "athleteCreationPage": {
       "firstName": "Vorname",
       "lastName": "Nachname",
       "E-Mail": "E-Mail",
       "birthdate": "Geburtstag",
-      "gender": "Geschlecht"
+      "gender": "Geschlecht",
+      "createButton": "erstelle Athlet*in"
     },
-
     "notFound": {
       "h1": "404",
       "h2": "Seite konnte nicht gefunden werden"
