+++ src/i18n/en-US/translation.json
@@ -32,15 +32,14 @@
       "header": "Athletes",
       "createButton": "create athlete"
     },
-
     "athleteCreationPage": {
       "firstName": "first name",
       "lastName": "last name",
-      "E-Mail": "E-Mail",
+      "E-Mail": "email",
       "birthdate": "birthdate",
-      "gender": "gender"
+      "gender": "gender",
+      "createButton": "create athlete"
     },
-
     "notFound": {
       "h1": "404",
       "h2": "Page not found"
