+++ src/pages/Athletes/AthleteOverviewPage.tsx
@@ -1,15 +1,13 @@
 import AthleteDatagrid from "@components/Datagrids/AthleteDatagrid/AthleteDatagrid";
-import { Box, Button, Modal, ModalClose, Sheet, Typography } from "@mui/joy";
+import { Box, Typography } from "@mui/joy";
 import { useTypedSelector } from "@stores/rootReducer";
 import { useTranslation } from "react-i18next";
-import AthleteCreationForm from "./AthleteCreationPage";
-import React, { useState } from "react";
+import AthleteCreationForm from "../../components/AthleteCreationPopup";
 
 const AthleteOverviewPage = () => {
   const athletes = useTypedSelector((state) => state.athletes.data);
   const athleteState = useTypedSelector((state) => state.athletes.state);
   const { t } = useTranslation();
-  const [open, setOpen] = useState(Boolean);
 
   return (
     <>
@@ -27,44 +25,7 @@ const AthleteOverviewPage = () => {
         <Typography level="h2" component="h1">
           {t("pages.athleteOverviewPage.header")}
         </Typography>
-
-        <React.Fragment>
-          <Button
-            variant="outlined"
-            color="neutral"
-            onClick={() => setOpen(true)}
-          >
-            {t("pages.athleteOverviewPage.createButton")}
-          </Button>
-          <Modal
-            aria-labelledby="modal-title"
-            aria-describedby="modal-desc"
-            open={open}
-            onClose={() => setOpen(false)}
-            sx={{
-              display: "flex",
-              justifyContent: "center",
-              alignItems: "center",
-            }}
-          >
-            <Sheet
-              variant="outlined"
-              sx={{ maxWidth: 1000, borderRadius: "md", p: 3, boxShadow: "lg" }}
-            >
-              <ModalClose variant="plain" sx={{ m: 1 }} />
-              <Typography
-                component="h2"
-                id="modal-title"
-                level="h4"
-                textColor="inherit"
-                sx={{ fontWeight: "lg", mb: 1 }}
-              ></Typography>
-              <Typography id="modal-desc" textColor="text.tertiary">
-                <AthleteCreationForm></AthleteCreationForm>
-              </Typography>
-            </Sheet>
-          </Modal>
-        </React.Fragment>
+        <AthleteCreationForm></AthleteCreationForm>
       </Box>
       <AthleteDatagrid
         athletes={athletes}
